refactor: use NetworkHandle to access num connected peers
